backend/requirements.txt
Flask
openai
pytesseract
Pillow
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dotenv
flask_cors
opencv-python
googletrans
pdf2image
flask_sqlalchemy
flask_bcrypt
sqlalchemy
psycopg2-binary
